% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_files}
\alias{all_files}
\title{Walkthrough 04 - Students with Disabilities Counts - Combined List}
\format{
Dataset of class \code{list}.
}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
all_files
}
\description{
Walkthrough 04 - List containing students with disabilities
longitudinal enrollment counts from 2012-18
}
\examples{

str(dataedu::all_files)
}
\keyword{datasets}
