% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{all_packages}
\alias{all_packages}
\title{All packages used in Data Science in Education Using R}
\format{A character vector}
\usage{
all_packages
}
\description{
All packages used in Data Science in Education Using R
}
\keyword{datasets}
