% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bchildcountandedenvironments2012}
\alias{bchildcountandedenvironments2012}
\title{Walkthrough 04 - Students with Disabilities Counts - 2012}
\format{
Dataset of class \code{tibble}.

Dataset of class \code{tibble}.
}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}

US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
bchildcountandedenvironments2012

bchildcountandedenvironments2012
}
\description{
Walkthrough 04 - Dataset containing students with disabilities
longitudinal enrollment counts in 2012

Walkthrough 04 - Dataset containing students with disabilities
longitudinal enrollment counts in 2012
}
\examples{

str(dataedu::bchildcountandedenvironments2012)

str(dataedu::bchildcountandedenvironments2012)
}
\keyword{datasets}
