% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bchildcountandedenvironments2013}
\alias{bchildcountandedenvironments2013}
\title{Walkthrough 04 - Students with Disabilities Counts - 2013}
\format{
Dataset of class \code{tibble}.
}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
bchildcountandedenvironments2013
}
\description{
Walkthrough 04 - Dataset containing students with disabilities
longitudinal enrollment counts in 2013
}
\examples{

str(dataedu::bchildcountandedenvironments2013)
}
\keyword{datasets}
