% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bluesky_posts}
\alias{bluesky_posts}
\title{Walkthrough 06 - Bluesky Data - Raw Posts}
\format{
Dataset of class \code{tibble}.
}
\source{
Bluesky API
}
\usage{
bluesky_posts
}
\description{
Walkthrough 06 - Raw data accessed from Bluesky on #tidytuesday
}
\examples{

str(dataedu::child_counts)
}
\keyword{datasets}
