% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{child_counts}
\alias{child_counts}
\title{Walkthrough 04 - Students with Disabilities Longitudinal Data - Combined Data Frame}
\format{
Dataset of class \code{tibble}.
}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
child_counts
}
\description{
Walkthrough 04 - Combined data frame containing students with disabilities
longitudinal enrollment counts from 2012-18
}
\examples{

str(dataedu::child_counts)
}
\keyword{datasets}
