% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dataedu-package.R
\docType{data}
\name{course_data}
\alias{course_data}
\title{Dataset on points earned out of points possible on more than 29,000 class assignments}
\format{Tibble
\describe{
}}
\source{
John Ranellucci and Joshua Rosenberg

John Ranellucci and Joshua Rosenberg
}
\usage{
course_data
}
\description{
Dataset on points earned out of points possible
on more than 29,000 class assignments
}
\examples{

str(dataedu::course_data)

}
\keyword{datasets}
