% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{course_minutes}
\alias{course_minutes}
\title{Dataset on log-trace data from the number of minutes students spend on the course learning management system for online science courses}
\format{Tibble
\describe{
}}
\source{
John Ranellucci and Joshua Rosenberg
}
\usage{
course_minutes
}
\description{
Dataset on log-trace data from the number of minutes students spend on the course learning management system for online science courses
}
\keyword{datasets}
