% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{package}
\name{dataedu-package}
\alias{dataedu}
\alias{dataedu-package}
\title{dataedu: Package for Data Science in Education Using R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provide readers of Data Science in Education Using
    R with a package containing useful functions, data, and references
    from the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-edu/dataedu}
  \item Report bugs at \url{https://github.com/data-edu/dataedu/issues}
}

}
\author{
\strong{Maintainer}: Isabella Velásquez \email{ivelasq@gmail.com}

Authors:
\itemize{
  \item Ryan Estrellado \email{ryan.estrellado@gmail.com}
  \item Emily Bovee \email{emily.a.bovee@gmail.com}
  \item Jesse Mostipak \email{jesse.maegan@gmail.com}
  \item Joshua Rosenberg \email{jmichaelrosenberg@gmail.com}
}

}
\keyword{internal}
