% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu.R
\docType{package}
\name{dataedu}
\alias{dataedu}
\title{dataedu: A package for Data Science in Education Using R}
\description{
The dataedu package provides readers of Data Science in Education Using R 
with a package with useful functions, data, and references from the book.
}
