% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataedu_bookdown_dependencies}
\alias{dataedu_bookdown_dependencies}
\title{All packages used to render the Data Science in Education Using R book (via {bookdown})}
\format{
A character vector
}
\usage{
dataedu_bookdown_dependencies
}
\description{
All packages used to render the Data Science in Education Using R book (via {bookdown})
}
\keyword{datasets}
