% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_dataedu.R
\docType{data}
\name{dataedu_colors}
\alias{dataedu_colors}
\title{Create a palette for Data Science in Education}
\format{An object of class \code{character} of length 5.}
\usage{
dataedu_colors
}
\description{
Create a palette for Data Science in Education
}
\keyword{datasets}
