% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_dataedu.R
\name{dataedu_cols}
\alias{dataedu_cols}
\title{Function to extract dataedu colors as hex codes}
\usage{
dataedu_cols(...)
}
\arguments{
\item{...}{Character names of dataedu_colors}
}
\description{
Function to extract dataedu colors as hex codes
}
