% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataedu_packages}
\alias{dataedu_packages}
\title{All packages used in Data Science in Education Using R}
\format{A character vector}
\usage{
dataedu_packages
}
\description{
All packages used in Data Science in Education Using R
}
\keyword{datasets}
