% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_dataedu.R
\name{dataedu_pal}
\alias{dataedu_pal}
\title{Return function to interpolate a dataedu color palette}
\usage{
dataedu_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dataedu_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a dataedu color palette
}
