% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{district_merged_data}
\alias{district_merged_data}
\title{Walkthrough 03 - Merged Ethnicity and FRPL District Data}
\format{Dataset of class \code{tibble}.}
\source{
Minneapolis Public Schools \url{https://mpls.k12.mn.us/}
}
\usage{
district_merged_data
}
\description{
Walkthrough 03 - Dataset with processed district data with
both FRPL and ethnicity data
}
\examples{

str(dataedu::district_merged_data)

}
\keyword{datasets}
