% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frpl_pdf}
\alias{frpl_pdf}
\title{Walkthrough 03 - Tabulizer Output from FRPL PDF}
\format{
Dataset of class \code{list}.
}
\source{
Minneapolis Public Schools \url{https://mpls.k12.mn.us/}
}
\usage{
frpl_pdf
}
\description{
Walkthrough 03 - RDS file with tabulizer output
from the FRPL PDF.
}
\examples{

str(dataedu::frpl_pdf)
}
\keyword{datasets}
