% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_dataedu.R
\docType{data}
\name{gillproto}
\alias{gillproto}
\title{Font for Data Science in Education}
\format{
An object of class \code{proto} (inherits from \code{environment}) of length 5.
}
\usage{
gillproto
}
\description{
Font for Data Science in Education
}
\keyword{datasets}
