% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dataedu.R
\name{install_bookdown_dependencies}
\alias{install_bookdown_dependencies}
\title{Mass installation of DSIEUR bookdown packages needed to render the book}
\usage{
install_bookdown_dependencies()
}
\description{
\code{install_bookdown_dependencies()} installs the packages required as
dependencies used to render the bookdown book
(at https://github.com/data-edu/data-science-in-education)
}
