% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_dataedu.R
\name{load.gillsansmt}
\alias{load.gillsansmt}
\title{load.gillsansmt}
\usage{
load.gillsansmt(font = "GillSansMT.TTF")
}
\arguments{
\item{font}{one of output from list.gillsansmt}
}
\description{
Load Gill Sans MT Font
}
