% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal_list}
\alias{longitudinal_list}
\title{List of datasets containing students with disabilities enrollment}
\format{}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
longitudinal_list
}
\description{
List of datasets containing students with disabilities enrollment
}
\keyword{datasets}
