% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_data_init}
\alias{ma_data_init}
\title{Foundational Skills Data}
\format{
Dataset of class \code{tibble}.
}
\usage{
ma_data_init
}
\description{
Dataset on Massachusetts public schools
}
\examples{

str(dataedu::ma_data_init)
}
\keyword{datasets}
