% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oulad_assessments}
\alias{oulad_assessments}
\title{Walkthrough 08 - Open University Learning Analytics Dataset - Assessment Data}
\format{
Dataset of class \code{tibble}.
}
\source{
Open University Learning Analytics Dataset \url{https://analyse.kmi.open.ac.uk/open-dataset}
}
\usage{
oulad_assessments
}
\description{
Walkthrough 08 - Provides data on students' performance on various assessments
throughout their courses.
}
\examples{

str(dataedu::oulad_assessments)
}
\keyword{datasets}
