% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oulad_interactions_filtered}
\alias{oulad_interactions_filtered}
\title{Walkthrough 08 - Open University Learning Analytics Dataset - Virtual Learning Environment}
\format{
Dataset of class \code{tibble}.
}
\source{
Open University Learning Analytics Dataset \url{https://analyse.kmi.open.ac.uk/open-dataset}
}
\usage{
oulad_interactions_filtered
}
\description{
Walkthrough 08 - Virtual learning environment (VLE)
log-trace data.
}
\examples{

str(dataedu::oulad_interactions_filtered)
}
\keyword{datasets}
