% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oulad_students}
\alias{oulad_students}
\title{Walkthrough 08 - Open University Learning Analytics Dataset - Student Data}
\format{
Dataset of class \code{tibble}.
}
\source{
Open University Learning Analytics Dataset \url{https://analyse.kmi.open.ac.uk/open-dataset}
}
\usage{
oulad_students
}
\description{
Walkthrough 08 - Information from three sources that relate to students and
the courses they took: \code{studentInfo}, \code{courses}, and \code{studentRegistration}.
}
\examples{

str(dataedu::oulad_students)
}
\keyword{datasets}
