% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dataedu-package.R
\docType{data}
\name{pre_survey}
\alias{pre_survey}
\title{Dataset on students' motivation administered before the start of those students' online science courses}
\format{Tibble
\describe{
}}
\source{
John Ranellucci and Joshua Rosenberg

John Ranellucci and Joshua Rosenberg
}
\usage{
pre_survey
}
\description{
Walkthrough 01 - Dataset on students' motivation administered before
the start of those students' online science courses
}
\examples{

str(dataedu::pre_survey)

}
\keyword{datasets}
