% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_pdf}
\alias{race_pdf}
\title{Walkthrough 03 - Tabulizer Output from Race PDF}
\format{
Dataset of class \code{list}.
}
\source{
Minneapolis Public Schools \url{https://mpls.k12.mn.us/}
}
\usage{
race_pdf
}
\description{
Walkthrough 03 - RDS file with tabulizer output
from the Race PDF.
}
\examples{

str(dataedu::race_pdf)
}
\keyword{datasets}
