% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sci_mo_with_text}
\alias{sci_mo_with_text}
\title{Walkthrough 01 - Student Motivation (Processed and With Text)}
\format{Dataset of class \code{tibble}.}
\source{
John Ranellucci and Joshua Rosenberg
}
\usage{
sci_mo_with_text
}
\description{
Walkthrough 01 - Processed dataset (for modeling with random forests)
from students on their motivation and achievement in
online science classes.
}
\examples{

str(dataedu::sci_mo_with_text)
}
\keyword{datasets}
