% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dataedu.R
\name{theme_dataedu}
\alias{theme_dataedu}
\title{ggplot2 theme for Data Science in Education Using R}
\usage{
theme_dataedu()
}
\description{
Apply the 'Data Science in Education Using R' background and font to ggplot2.
}
\details{
The dataedu theme uses the font Cabin.
}
