% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{tt_tweets}
\alias{tt_tweets}
\title{Walkthrough 12 - Tweet Data}
\format{Dataset of class \code{tibble}.}
\source{
Twitter and TidyTuesday \url{https://thomasmock.netlify.com/post/tidytuesday-a-weekly-social-data-project-in-r/}
}
\description{
Tweet dataset gathered from hashtag tidytuesday
}
\keyword{#'}
\keyword{@examples}
\keyword{datasets}
\keyword{str(dataedu::tt_tweets)}
