% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{course_data}
\alias{course_data}
\title{Walkthrough 01 - Course Data}
\format{Dataset of class \code{tibble}.}
\source{
John Ranellucci and Joshua Rosenberg
}
\usage{
course_data
}
\description{
Dataset on points earned out of points possible
on more than 29,000 class assignments
}
\examples{

str(dataedu::course_data)
}
\keyword{datasets}
