% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{course_minutes}
\alias{course_minutes}
\title{Walkthrough 01 - Course Minutes}
\format{Dataset of class \code{tibble}.}
\source{
John Ranellucci and Joshua Rosenberg
}
\description{
Walkthrough 01 - Dataset on log-trace data from the number of minutes
students spend on the course learning management system
for online science courses
}
\examples{

str(dataedu::course_minutes)

}
\keyword{datasets}
