% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{package}
\name{dataedu-package}
\alias{dataedu}
\alias{dataedu-package}
\title{dataedu: Package for Data Science in Education Using R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provide readers of Data Science in Education Using R with 
    a package with useful functions, data, and references from the book.
}
\author{
\strong{Maintainer}: Isabella Velasquez \email{ivelasq@gmail.com}

Authors:
\itemize{
  \item Emily Bovee \email{emily.a.bovee@gmail.com}
  \item Ryan Estrellado \email{ryan.estrellado@gmail.com}
  \item Jesse Mostipak
  \item Joshua Rosenberg \email{jmichaelrosenberg@gmail.com}
}

}
\keyword{internal}
