% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal_data}
\alias{longitudinal_data}
\title{Datasets containing students with disabilities enrollment}
\format{Tibble
\describe{
}}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
longitudinal_data
}
\description{
Datasets containing students with disabilities enrollment
}
\keyword{datasets}
