% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal_data}
\alias{longitudinal_data}
\title{Walkthrough 04 - Students with Disabilities Longitudinal Data}
\format{Dataset of class \code{tibble}.}
\source{
US Department of Education \url{https://www2.ed.gov/programs/osepidea/618-data/state-level-data-files/index.html}
}
\usage{
longitudinal_data
}
\description{
Walkthrough 04 - Datasets containing students with disabilities
longitudinal enrollment
}
\examples{

str(dataedu::longitudinal_data)

}
\keyword{datasets}
