% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{sci_mo_data}
\alias{sci_mo_data}
\title{Walkthrough 01 - Student Motivation}
\format{Dataset of class \code{tibble}.}
\source{
John Ranellucci and Joshua Rosenberg
}
\description{
Dataset from students on their motivation and achievement
in online science classes
}
\examples{

str(dataedu::sci_mo_data)

\describe{
}

}
\keyword{datasets}
