% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sci_mo_processed}
\alias{sci_mo_processed}
\title{Processed dataset (for modeling with multi-level models) from students on their motivation and achievement in online science classes}
\format{Tibble
\describe{
}}
\source{
John Ranellucci and Joshua Rosenberg
}
\usage{
sci_mo_processed
}
\description{
Processed dataset (for modeling with multi-level models) from students on their motivation and achievement in online science classes
}
\keyword{datasets}
