% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dataedu-package.R
\docType{data}
\name{sci_mo_processed}
\alias{sci_mo_processed}
\title{Processed dataset (for modeling with multi-level models) from students on their motivation and achievement in online science classes}
\format{Tibble
\describe{
}}
\source{
John Ranellucci and Joshua Rosenberg

John Ranellucci and Joshua Rosenberg
}
\usage{
sci_mo_processed
}
\description{
Walkthrough 01 - Processed dataset (for modeling with multi-level models)
from students on their motivation and achievement in
online science classes
}
\examples{

str(dataedu::sci_mo_processed)

}
\keyword{datasets}
