% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tt_tweets}
\alias{tt_tweets}
\title{Walkthrough 12 - Tweet Data}
\format{Dataset of class \code{tibble}.}
\source{
Twitter and TidyTuesday \url{https://thomasmock.netlify.com/post/tidytuesday-a-weekly-social-data-project-in-r/}
}
\usage{
tt_tweets
}
\description{
Walkthrough 12 - Tweet dataset gathered from hashtag tidytuesday
}
\examples{

str(dataedu::tt_tweets)
}
\keyword{datasets}
