% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataedu-package.R
\docType{data}
\name{tt_tweets}
\alias{tt_tweets}
\title{Walkthrough 12 - Tweet Data}
\format{A data frame with columns:
\describe{
\item{status_id}{A unique identifier for a tweet}
\item{lang}{Language of a tweet}
\item{text}{Text of a tweet}
}}
\source{
Twitter and TidyTuesday \url{https://thomasmock.netlify.com/post/tidytuesday-a-weekly-social-data-project-in-r/}
}
\description{
Tweet dataset gathered from hashtag tidytuesday
}
\keyword{#'}
\keyword{@examples}
\keyword{datasets}
\keyword{str(dataedu::tt_tweets)}
